chore: refactor verification context
